# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRMIS API Documentation",
        "description": "Human Resource Management Information System API",
        "contact": {
            "name": "HRMIS Support",
            "email": "admin@hrmis.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/tasks": {
            "get": {
                "tags": [
                    "Task Management"
                ],
                "summary": "Get all tasks",
                "description": "Retrieve a list of all tasks with assigned user information (Admin/Manager access required)",
                "operationId": "getTasks",
                "responses": {
                    "200": {
                        "description": "List of tasks retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Complete project documentation"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Write comprehensive documentation for the new feature"
                                            },
                                            "priority": {
                                                "type": "string",
                                                "enum": [
                                                    "Low",
                                                    "Medium",
                                                    "High",
                                                    "Critical"
                                                ],
                                                "example": "High"
                                            },
                                            "status": {
                                                "type": "string",
                                                "enum": [
                                                    "Pending",
                                                    "In Progress",
                                                    "Completed",
                                                    "On Hold"
                                                ],
                                                "example": "In Progress"
                                            },
                                            "assigned_to": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "assigned_to_user": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "john@example.com"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing authentication token"
                    },
                    "403": {
                        "description": "Forbidden - Insufficient permissions (Admin/Manager role required)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Task Management"
                ],
                "summary": "Create a new task",
                "description": "Create a new task and assign it to an employee (Admin/Manager access required)",
                "operationId": "createTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "priority",
                                    "status",
                                    "assigned_to"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Task title",
                                        "type": "string",
                                        "example": "Complete project documentation"
                                    },
                                    "description": {
                                        "description": "Detailed task description",
                                        "type": "string",
                                        "example": "Write comprehensive documentation for the new feature"
                                    },
                                    "priority": {
                                        "description": "Task priority level",
                                        "type": "string",
                                        "enum": [
                                            "Low",
                                            "Medium",
                                            "High",
                                            "Critical"
                                        ],
                                        "example": "High"
                                    },
                                    "status": {
                                        "description": "Current task status",
                                        "type": "string",
                                        "enum": [
                                            "Pending",
                                            "In Progress",
                                            "Completed",
                                            "On Hold"
                                        ],
                                        "example": "Pending"
                                    },
                                    "assigned_to": {
                                        "description": "ID of the user assigned to the task",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Complete project documentation"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Write comprehensive documentation for the new feature"
                                        },
                                        "priority": {
                                            "type": "string",
                                            "example": "High"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Pending"
                                        },
                                        "assigned_to": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing authentication token"
                    },
                    "403": {
                        "description": "Forbidden - Insufficient permissions (Admin/Manager role required)"
                    },
                    "422": {
                        "description": "Validation error - Invalid input data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}